# generated/src/main/java/com/shoppingcart/user/api/RecoveryPasswordApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.shoppingcart.user.api;

import com.shoppingcart.user.model.RecoveryPasswordRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "public", description = "Public Access API  Method")
public interface RecoveryPasswordApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /recovery-password : Recovery password
     * Recovery password 
     *
     * @param recoveryPasswordRequest  (optional)
     * @return success Email Send (status code 200)
     *         or bad input parameter (status code 400)
     */
    @Operation(
        operationId = "recoverPassword",
        summary = "Recovery password",
        description = "Recovery password ",
        tags = { "public" },
        responses = {
            @ApiResponse(responseCode = "200", description = "success Email Send"),
            @ApiResponse(responseCode = "400", description = "bad input parameter")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/recovery-password",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> recoverPassword(
        @Parameter(name = "RecoveryPasswordRequest", description = "") @Valid @RequestBody(required = false) RecoveryPasswordRequest recoveryPasswordRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}